requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
pillow==12.1.0
python-docx==1.2.0
python-pptx==1.0.2
typing_extensions==4.15.0
Werkzeug==3.1.4
xlsxwriter==3.2.9

